feat: integrate map component into current layout
