--- src/components/Main/Map/map.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './map.scss';
-
-const Map = () => (
-  <section className="map">
-    <div className="">
-      map is here
-    </div>
-  </section>
-);
-
-export default Map;
