--- src/components/Map/index.scss
@@ -1,9 +1,0 @@
-.map {
-  margin: 1rem;
-  padding: 1rem;
-}
-
-.mapArea {
-  border: solid 0.1rem white;
-  border-radius: 1rem;
-}
